requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
bitarray==3.0.0
blessed==1.20.0
bravado==10.4.1
bravado-core==6.1.1
Bybit==0.2.12
ccxt==4.4.44
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
coinbase==2.1.0
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==44.0.0
cytoolz==1.0.1
distlib==0.3.9
Django==4.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-decouple==2.1
django-dotenv==1.4.2
django-environ==0.11.2
django-picklefield==3.2
django-q==1.3.9
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
filelock==3.16.1
fqdn==1.5.1
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
importlib_resources==6.4.5
isoduration==20.11.0
jinxed==1.3.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kucoin-python==1.0.25
MetaTrader5==5.0.4738
monotonic==1.6
msgpack==1.1.0
msgpack-python==0.5.6
multidict==6.1.0
numpy==2.2.1
packaging==24.2
parsimonious==0.10.0
paypalrestsdk==1.13.3
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
propcache==0.2.1
pybit==5.8.0
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pyOpenSSL==24.3.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
pywin32==308
PyYAML==6.0.2
qrcode==8.0
redis==3.5.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==4.0.1
rpds-py==0.22.3
setuptools==75.6.0
shortuuid==1.0.13
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
swagger-spec-validator==3.0.4
toolz==1.0.0
types-python-dateutil==2.9.0.20241206
types-requests==2.31.0.20240406
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.15
virtualenv==20.28.0
wcwidth==0.2.13
web3==7.6.1
webcolors==24.11.1
websocket-client==1.8.0
websockets==13.1
yarl==1.18.3
